# Implement Liquid::Raw#parse