Implement `Liquid::Raw#parse`.

The function `match_full_token_possibly_invalid` emulates matching the `FullTokenPossiblyInvalid = /\A(.*)#{TagStart}\s*(\w+)\s*(.*)?#{TagEnd}\z/om` regex.

# Benchmark

For the benchmark, I wrapped every `theme.liquid` in a `raw` tag and ran the benchmark.

Master:

```
              parse:    175.223  (± 9.7%) i/s -      1.746k in  10.078290s
             render:    250.375  (± 8.4%) i/s -      2.496k in  10.042677s
     parse & render:     96.516  (± 7.3%) i/s -    960.000  in  10.009864s
```

This branch:

```
              parse:    280.836  (± 9.3%) i/s -      2.793k in  10.040895s
             render:    236.161  (±12.3%) i/s -      2.340k in  10.076232s
     parse & render:    120.075  (± 9.2%) i/s -      1.199k in  10.070105s
```

In this (very) synthetic benchmark, we see a 60% increase in performance during parsing.